+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 ==================
 
 ## 0.18.3
+### is.codion.swing.common.ui
+- Controls.Config.standard() renamed control().
 
 ## 0.18.2
 ### is.codion.swing.common.model